# docs/contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ExtensionSuite.Host | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/recovery/restore": {
      "post": {
        "tags": [
          "ExtensionSuite.Host"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/integrations/streamerbot/actions/execute": {
      "post": {
        "tags": [
          "ExtensionSuite.Host"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotExecution"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/streamerbot/triggers/execute": {
      "post": {
        "tags": [
          "ExtensionSuite.Host"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeTriggerExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotExecution"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/integrations/speakerbot/speak": {
      "post": {
        "tags": [
          "ExtensionSuite.Host"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/integrations/speakerbot/queue": {
      "post": {
        "tags": [
          "ExtensionSuite.Host"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeakerQueueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/inspector": {
      "get": {
        "tags": [
          "ExtensionSuite.Host"
        ],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InspectorEntry"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/inspector/{id}/fixture": {
      "get": {
        "tags": [
          "ExtensionSuite.Host"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/inspector/{id}/replay": {
      "post": {
        "tags": [
          "ExtensionSuite.Host"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectorReplayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionExecutionRequest": {
        "required": [
          "actionId"
        ],
        "type": "object",
        "properties": {
          "actionId": {
            "type": "string",
            "format": "uuid"
          },
          "arguments": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonObject"
              }
            ]
          },
          "executeLive": {
            "type": "boolean",
            "default": false
          }
        },
        "additionalProperties": false
      },
...
      "BotExecution": {
        "required": [
          "id",
          "operation",
          "state",
          "mayHaveExecuted",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "operation": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "mayHaveExecuted": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "CanonicalEvent": {
        "required": [
          "source",
          "platform",
          "type",
          "nativeType",
          "dedupeKey"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "source": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "nativeType": {
            "type": "string"
          },
          "nativeId": {
            "type": [
              "null",
              "string"
            ]
          },
          "user": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventUser"
              }
            ]
          },
          "message": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventMessage"
              }
            ]
          },
          "monetary": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventMoney"
              }
            ]
          },
          "support": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SupportDetails"
              }
            ]
          },
          "stream": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventStream"
              }
            ]
          },
          "metrics": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventMetrics"
              }
            ]
          },
          "automation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventAutomationMetadata"
              }
            ]
          },
          "dedupeKey": {
            "type": "string"
          },
          "raw": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonObject"
              }
            ]
          },
          "provenance": {
            "$ref": "#/components/schemas/EventProvenance"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "bridgePath": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "CodeTriggerExecutionRequest": {
        "required": [
          "eventName"
        ],
        "type": "object",
        "properties": {
          "eventName": {
            "type": "string"
          },
          "arguments": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonObject"
              }
            ]
          },
          "executeLive": {
            "type": "boolean",
            "default": false
          }
        },
        "additionalProperties": false
      },
...
      "EventAutomationMetadata": {
        "type": "object",
        "properties": {
          "anonymous": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "messagePublic": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "language": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
      "EventBadge": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "version": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
      "EventMessage": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": [
              "null",
              "string"
            ]
          },
          "parts": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/EventMessagePart"
            }
          }
        },
        "additionalProperties": false
      },
      "EventMessagePart": {
        "required": [
          "kind",
          "text"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "source": {
            "type": [
              "null",
              "string"
            ]
          },
          "zeroWidth": {
            "type": "boolean",
            "default": false
          }
        },
        "additionalProperties": false
      },
      "EventMetrics": {
        "type": "object",
        "properties": {
          "value": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "health": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
      "EventMoney": {
        "required": [
          "minorUnits",
          "currency",
          "valuationKind"
        ],
        "type": "object",
        "properties": {
          "minorUnits": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "valuationKind": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "EventProvenance": {
        "enum": [
          "live",
          "simulation",
          "replay"
        ]
      },
      "EventStream": {
        "required": [
          "id",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
      "EventUser": {
        "type": "object",
        "properties": {
          "platformUserId": {
            "type": [
              "null",
              "string"
            ]
          },
          "login": {
            "type": [
              "null",
              "string"
            ]
          },
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "avatarUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "badges": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "isBot": {
            "type": "boolean",
            "default": false
          },
          "badgeDetails": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/EventBadge"
            }
          }
        },
        "additionalProperties": false
      },
...
      "InspectorEntry": {
        "required": [
          "id",
          "receivedAt",
          "classification",
          "limitation",
          "event",
          "payload"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "classification": {
            "type": "string"
          },
          "limitation": {
            "type": [
              "null",
              "string"
            ]
          },
          "event": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CanonicalEvent"
              }
            ]
          },
          "payload": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonObject"
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "InspectorReplayRequest": {
        "type": "object",
        "properties": {
          "persist": {
            "type": "boolean",
            "default": false
          }
        },
        "additionalProperties": false
      },
...
      "JsonObject": {
        "type": "object"
      },
...
      "NativeMoney": {
        "required": [
          "amountMinor",
          "currency",
          "minorUnitDigits"
        ],
        "type": "object",
        "properties": {
          "amountMinor": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "currency": {
            "type": "string"
          },
          "minorUnitDigits": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "RestoreRequest": {
        "required": [
          "id",
          "confirm"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "confirm": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SpeakerQueueRequest": {
        "required": [
          "operation"
        ],
        "type": "object",
        "properties": {
          "operation": {
            "type": "string"
          },
          "value": {
            "type": [
              "null",
              "string"
            ]
          },
          "executeLive": {
            "type": "boolean",
            "default": false
          }
        },
        "additionalProperties": false
      },
      "SpeechRequest": {
        "required": [
          "voice",
          "message"
        ],
        "type": "object",
        "properties": {
          "voice": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "badWordFilter": {
            "type": "boolean",
            "default": true
          },
          "executeLive": {
            "type": "boolean",
            "default": false
          }
        },
        "additionalProperties": false
      },
...
      "SupportDetails": {
        "required": [
          "kind",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "nativeMoney": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NativeMoney"
              }
            ]
          },
          "tier": {
            "type": "string",
            "default": ""
          },
          "giftCorrelationKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "giftRole": {
            "type": "string",
            "default": "none"
          },
          "gatedReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "reportedAmountMajor": {
            "type": [
              "null",
              "string"
            ]
          },
          "reportedCurrency": {
            "type": [
              "null",
              "string"
            ]
          },
          "giftRecipientKeys": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "giftScopeKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "giftPeriodStart": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "giftPeriodEnd": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}